Refactored project and added new Class for Calories Limits

+++ src/com/olehpodolin/model/OtherProducts.java
@@ -2,7 +2,7 @@ package com.olehpodolin.model;
 
 public class OtherProducts extends Product{
 
-    public OtherProducts(String name, float carbs, float fats, float protein) {
+    public OtherProducts(String name, double carbs, double fats, double protein) {
         super(name, carbs, fats, protein);
     }
 }
