+++ src/com/olehpodolin/model/RichInCarbsProduct.java
@@ -2,7 +2,7 @@ package com.olehpodolin.model;
 
 public class RichInCarbsProduct extends Product {
 
-    public RichInCarbsProduct(String name, float carbs, float fats, float protein) {
+    public RichInCarbsProduct(String name, double carbs, double fats, double protein) {
         super(name, carbs, fats, protein);
     }
 }
