+++ src/com/olehpodolin/model/RichInFatsProduct.java
@@ -2,7 +2,7 @@ package com.olehpodolin.model;
 
 public class RichInFatsProduct extends Product {
 
-    public RichInFatsProduct(String name, float carbs, float fats, float protein) {
+    public RichInFatsProduct(String name, double carbs, double fats, double protein) {
 
         super(name, carbs, fats, protein);
     }
