+++ src/com/olehpodolin/model/RichInProteinProduct.java
@@ -2,7 +2,7 @@ package com.olehpodolin.model;
 
 public class RichInProteinProduct extends Product{
 
-    public RichInProteinProduct(String name, float carbs, float fats, float protein) {
+    public RichInProteinProduct(String name, double carbs, double fats, double protein) {
         super(name, carbs, fats, protein);
     }
